Make a function for loading data and fix real data workflow
The issue was that the windows and positions weren't matching anymore
when the positions were excluded for z>0. I put the Euler Deconv and
inversion in the same loop to avoid this. I can now reproduce the
previous results.

But something seems off when plotting the inclination and declination
spatially. There are anomalies with the same polarity coming up with
opposite declinations. And some of the smaller anomalies are being
ignored completely when they look like perfectly fine dipoles as well.